// src/common/enums/post-category.enum.ts
export enum PostCategory {
    STUDY = 'study',
    TRAVEL = 'travel',
    CLASS_REUNION = 'class_reunion',
    SHOPPING = 'shopping',
    FOOD = 'food',
    SPORTS = 'sports',
    MUSIC = 'music',
    MOVIE = 'movie',
    GAMING = 'gaming',
    TECHNOLOGY = 'technology',
    WORK = 'work',
    FAMILY = 'family',
    FRIENDS = 'friends',
    LOVE = 'love',
    HEALTH = 'health',
    FASHION = 'fashion',
    PET = 'pet',
    NEWS = 'news',
    EVENT = 'event',
    FUNNY = 'funny',
    LIFE = 'life',
    OTHER = 'other',
}